+++ eyes_common/applitools/common/app_output.py
@@ -1,4 +1,4 @@
-from typing import Text, Optional
+from typing import Optional, Text
 
 import attr
 
@@ -10,9 +10,7 @@ __all__ = ("AppOutput",)
 @attr.s
 class AppOutput(object):
     title = attr.ib(metadata={JsonInclude.THIS: True})  # type: Text
-    screenshot64 = attr.ib(
-        repr=False, metadata={JsonInclude.NON_NONE: True}
-    )  # type: Optional[Text]
+    screenshot_bytes = attr.ib(repr=False)  # type: Optional[bytes]
     screenshot_url = attr.ib(
         default=None, metadata={JsonInclude.NON_NONE: True}
     )  # type: Optional[Text]